fix killer conspiracy double-bug in cross product
there were a pair of perfectly symmetrical bugs in
our definition of the cross product matrix and
in the tables to construct tetrahedra in a cube.
we were getting the right volumes, qualities, etc.,
but our determinant code is broken.
fix the two conspirators and add some documentation
about the tetrahedra tables