added random perturbations to RIB
if the points have no best axis
and the arbitrary axis chosen
can't differentiate the points
in the middle, we can perturb
the already arbitrary axis.
this makes the difference
for 4 perfectly spaced points.
if this fails, a warning is given
about points being the same.
this is likely to catch
meshes where all vertices are
at (0,0,0), for example.